Added block allocation strategy.
A single block allocates memory on the heap for objects with the same
size. When an object is between two block sizes, e.g. 256 and 512 bytes
the object is allocated on the larger block memory buffer which leads
to some memory being wasted.

Each block of memory contains smaller chunks that are used to track
what sections of memory are left. A chunk by default can hold 64
objects (independent of size) and keeps track of the available memory
slots using a bitmask (64 bits). Deallocation happens by finding the
appropriate block and then use that do find the chunk and then flipping
the corresponding bit.