Added initial support for modules at runtime.
This involved refactoring the current approach of one interpreter
per runtime. Instead, now every time a new module is imported, a
new interpreter is instantiated, which ensures that we don't start
accessing each modules namespace accidentally.

The VM manages the interpreters, through InterpreterSession, and
every time a new Program needs to be interpreted (for example when
a module is imported), it creates a new interpreter. When module
import is done, the calling interpreter copies the module's Interpreter
globals to the module's symbol table, and then shuts down the
Interpreter.