Removed PyObjectType enum, since it is no longer needed.
This enum was largely used to make sure that casting from PyObject*
to a derived type was safe. The PyObjectType enum overlaps with
in PyType in functionality. So, now casting to a derived PyObject*
involves checking that PyObject::type returns a pointer to the
PyType singleton we want to cast to. For example, casting `obj` (a
PyObject*) to PyTuple* is only valid if `obj->type() == tuple();`.