Added runtime support for expressions with `and` and/or `or` keywords.
Additionally, cleaned up the Label class. Now, a Label is a
std::shared_ptr, which is owned by the instructions that need it to
jump to (e.g. JumpIf*) and the generator holds a reference to the
Labels rather than storing them in std::set.

This also required adding parser support for groups in order to write
some unittests (e.g. (True and False) == False is not the same as
True and False == False, and has a different result).