Added the concept of function blocks.
Until now a function was a stream of instructions, but for try
statements it makes sense to have blocks within a function.
When an exception occurs in a try body we can then exit the block,
rather than exit the function (which would prevent us from running
the code in except and finally clauses, which may rescue execution).
For now each function has a single block which contains all
instructions.

Having blocks could also be useful for smarter register allocations
and other optimisations. In LLVM blocks are also the fundamental
instruction groups, which means someone smarter than me thinks this :)