Implemented logic to handle exception handlers.
This may be temporary. The current approach is to stash exceptions,
and clear the exception state of the current frame, and then compare
the stashed exception to each exception handler. Exceptions have to
be stashed (at least currently) otherwise we are in a error state and
cannot execute the code in the exception handler (which could be a
function call).