Added ast::Value to represent a value in the IR.
For example, in the BytecodeGenerator Value is mainly a wrapper around
Register. In LLVM it is a wrapper around llvm::Value.

Also added namespace py because the Value type was becoming confusing
since there is already a Value for the runtime, that represents any
possible value that is representable in Python.