Add objects to GC roots that are on the stack.
Some objects are anonymous, ie. the user cannot access them by name and
they are not part of any scope. For example, the range_iterator when
doing a ranged for loop is not associated to any variable. Therefore,
the VM must keep track of all objects on the stack, so the the GC can
add them to the root list.