Fixed issue where return was parsed as a variable name.
This was due to the fact that return was allowed to be parsed as a
variable name, and consequently something like `return [1]` was parsed
as slicing a variable called return.

To avoid this issue in the future all reserved keywords are stored in a
tuple and when matching strings against atom pattern, we first check if
a TokenType::NAME corresponds to a string in reserved keywords.