Runtime: Fixed Value equality comparisson in unordered containers
There was a bug where the Value equality in unordered containers was not
being performed using the ValueEqual call operator, but rather the std
variant operator==. This meant that for PyDict equal comparisson there
was UB, since Value operator== was not giving the same result as
ValueEqual. Now, instead of ValueEqual we just added operator== for
Value, with the same logic that ValueEqual had.