Runtime: Added PyObject::true_ to determined truthyness of an object
This fixes the previous behaviour where we had PyObject::__bool__
defaulted to be true. Now we check if __bool__ is implemented, then
check if the object implements length and finally if none of the above
was called we return true. PyNone is always false, PyBool returns true
or false depending on its value.