BytecodeGenerator+Runtime: Fixed cell2arg
cell2arg now stores the complete information about whether a given
argument is a cell variable. cell2arg is an array of size length of
cellvars, where each element corresponds to the argument index of a
cell, if it is an argument. If it isn't an argument the value at the
corresponding cell index in cell2arg is set to the total number of
arguments of the function (ie. it is not possible to index into the
arguments with this value since it is out of bounds).