Compiler: Fixed symbol scope resolution for implicit globals
When an identifier is defined in function scope, but has been previously
deduced as implicitly global (due to definition in global scope), it
should have local scope.

```python
a = 1
def foo():
    print(a)
    a = 2

foo()
```
This snippet raises an UnboundLocalError because `a` in `foo` is a local
symbol, but is not bound at `print(a)`.