MLIR: Initial MLIR backend implementation
Currently this brings the backend to parity with BytecodeGenerator.
This is done by lowerig AST to the Python dialect, which is then lowered
to emitpybytecode. The latter is a close mapping to the current
instructions supported by the python-cpp bytecode. In future iterations
the Python dialect could/should be lowered to builtin MLIR dialects,
which can then be lowered to LLVM and jitted at runtime (or something
along these lines).